I added an optional write so I could edit the code without writing out all the datasets.
I put together `RecObs` and added the edits I did with Danielle.  It was too hard to go back and do in in t`tblObservations`.

Doing the spatial intersect on StrPlot with RecObs only ended up with 137 observations, so a good number of the locations are incorrect.  Most of the strata and plots are though, so do the estimation.