Not sure what `inCNA` stands for, augmented data probably.  Not used though.
Made the augmented `FWSColonyID` unique.  The were 1:nAug.

Put Year, Str, in a mutate statement.

Corrected the rownames for `MltObsr` to `FWSColonyID`.

Removed ColObs because it is not used, or its indices.

Now keeping the model code in `extdata` itself.

Added `here::here` to the simulations and estiamtes files.

Added titles to the output tables in the spreadsheet and set the titles above their tables.