Put the meta-variable dump file in data not in ori.
Changed FS for year-frame-stratum to YFS.

Change cDFStr to cStr.  For the headings, add the Total in or make a new variable.

Add the observer meta-variables in the setup program here.

Need Danielle to track the observations that do not have a stratum or plot.

Changed the detection so the rear observer only has survey year estimates.  By observer and year did not have enough capture histories.

Removed the meta-variables from the save file because eventually, we will get everything from the CSV files in the data directory.